feat: ajustado para build en producción
