+++ package.json
@@ -5,7 +5,7 @@
     "scripts": {
         "dev": "next dev --turbopack -p 4000",
         "build": "next build",
-        "start": "next start",
+        "start": "next start -p 4000",
         "lint": "next lint"
     },
     "dependencies": {
