--- src/wasm/wasmd_html_bg.wasm.d.ts
@@ -1,9 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-export const memory: WebAssembly.Memory;
-export const markdown_to_html: (a: number, b: number) => [number, number];
-export const __wbindgen_export_0: WebAssembly.Table;
-export const __wbindgen_malloc: (a: number, b: number) => number;
-export const __wbindgen_realloc: (a: number, b: number, c: number, d: number) => number;
-export const __wbindgen_free: (a: number, b: number, c: number) => void;
-export const __wbindgen_start: () => void;
